+++ test/expectation_values.jl
@@ -4,7 +4,7 @@ using Test
 @testset "Expectation values" begin
     @testset "Qubits (d=2)" begin
         # Qubit |0⟩ stabilized by Z.
-        tab = reshape(Int64[0, 1, 0], 3, 1)
+        tab = reshape(Int[0, 1, 0], 3, 1)
         stab = StabilizerTableau(2, 1, tab; m=1, storephase=true)
 
         # Identity always has expectation 1.
@@ -24,7 +24,7 @@ using Test
         @test QuditClifford.expect_int!(stab, Int[0, 1, 2]) == 2
 
         # If phases are not stored, in-span operators still return 1 by convention.
-        tab_nophase = reshape(Int64[0, 1], 2, 1)
+        tab_nophase = reshape(Int[0, 1], 2, 1)
         stab_nophase = StabilizerTableau(2, 1, tab_nophase; m=1, storephase=false)
         @test QuditClifford.expect!(stab_nophase, Int[0, 1]) ≈ 1.0 + 0.0im
         @test QuditClifford.expect_int!(stab_nophase, Int[0, 1]) == 0
@@ -42,7 +42,7 @@ using Test
 
     @testset "Qudits (d=3)" begin
         # Qutrit |0⟩ stabilized by Z.
-        tab = reshape(Int64[0, 1, 0], 3, 1)
+        tab = reshape(Int[0, 1, 0], 3, 1)
         stab = StabilizerTableau(3, 1, tab; m=1, storephase=true)
 
         # Identity always has expectation 1.
@@ -63,7 +63,7 @@ using Test
         @test QuditClifford.expect_int!(stab, Int[0, 1, 1]) == 1
 
         # If phases are not stored, in-span operators still return 1 by convention.
-        tab_nophase = reshape(Int64[0, 1], 2, 1)
+        tab_nophase = reshape(Int[0, 1], 2, 1)
         stab_nophase = StabilizerTableau(3, 1, tab_nophase; m=1, storephase=false)
         @test QuditClifford.expect!(stab_nophase, Int[0, 1]) ≈ 1.0 + 0.0im
         @test QuditClifford.expect_int!(stab_nophase, Int[0, 1]) == 0
