+++ test/measurements.jl
@@ -12,7 +12,7 @@ using Random
         # and set the stabilizer phase to match the sampled outcome.
         outcomes = Int[]
         for _ in 1:10
-            tab = reshape(Int64[0, 1, 0], 3, 1)
+            tab = reshape(Int[0, 1, 0], 3, 1)
             stab = StabilizerTableau(2, 1, tab; m=1, storephase=true)
 
             out = measure!(stab, Int[1, 0]) # measure X
@@ -26,7 +26,7 @@ using Random
         @test sort!(unique(outcomes)) == [0, 1]  # ensure we got both outcomes at least once
 
         # Commuting measurement in span (Z) should be deterministic and leave the state.
-        tab_det = reshape(Int64[0, 1, 0], 3, 1)
+        tab_det = reshape(Int[0, 1, 0], 3, 1)
         stab_det = StabilizerTableau(2, 1, tab_det; m=1, storephase=true)
         out_det = measure!(stab_det, Int[0, 1]) # measure Z
         @test out_det == 0
@@ -37,7 +37,7 @@ using Random
         # Maximally mixed qubit (m=0): commuting but not in span should append a generator.
         outcomes2 = Int[]
         for _ in 1:10
-            tab2 = zeros(Int64, 3, 1)
+            tab2 = zeros(Int, 3, 1)
             stab2 = StabilizerTableau(2, 1, tab2; m=0, storephase=true)
 
             out2 = measure!(stab2, Int[0, 1]) # measure Z
@@ -60,7 +60,7 @@ using Random
         # and set the stabilizer phase to match the sampled outcome.
         outcomes = Int[]
         for _ in 1:20
-            tab = reshape(Int64[0, 1, 0], 3, 1)
+            tab = reshape(Int[0, 1, 0], 3, 1)
             stab = StabilizerTableau(3, 1, tab; m=1, storephase=true)
 
             out = measure!(stab, Int[1, 0]) # measure X
@@ -74,7 +74,7 @@ using Random
         @test sort!(unique(outcomes)) == [0, 1, 2]  # ensure we got all outcomes at least once
 
         # Commuting measurement in span (Z) should be deterministic and leave the state.
-        tab_det = reshape(Int64[0, 1, 0], 3, 1)
+        tab_det = reshape(Int[0, 1, 0], 3, 1)
         stab_det = StabilizerTableau(3, 1, tab_det; m=1, storephase=true)
         out_det = measure!(stab_det, Int[0, 1]) # measure Z
         @test out_det == 0
@@ -85,7 +85,7 @@ using Random
         # Maximally mixed qutrit (m=0): commuting but not in span should append a generator.
         outcomes2 = Int[]
         for _ in 1:20
-            tab2 = zeros(Int64, 3, 1)
+            tab2 = zeros(Int, 3, 1)
             stab2 = StabilizerTableau(3, 1, tab2; m=0, storephase=true)
 
             out2 = measure!(stab2, Int[0, 1]) # measure Z
